move single use utility functions
