+++ seqr/views/utils/variant_utils.py
@@ -3,7 +3,6 @@ from django.contrib.postgres.aggregates import ArrayAgg
 from django.db.models import F
 import logging
 import redis
-from typing import Callable
 
 from matchmaker.models import MatchmakerSubmissionGenes, MatchmakerSubmission
 from reference_data.models import TranscriptInfo
@@ -307,10 +306,6 @@ def get_variant_main_transcript(variant):
     return {}
 
 
-def get_sv_name(variant_json):
-    return variant_json.get('svName') or '{svType}:chr{chrom}:{pos}-{end}'.format(**variant_json)
-
-
 def get_saved_discovery_variants_by_family(variant_filter, format_variants, get_family_id):
     tag_types = VariantTagType.objects.filter(project__isnull=True, category=DISCOVERY_CATEGORY)
 
@@ -327,92 +322,3 @@ def get_saved_discovery_variants_by_family(variant_filter, format_variants, get_
         saved_variants_by_family[family_id].append(saved_variant)
 
     return saved_variants_by_family
-
-
-HET = 'Heterozygous'
-HOM_ALT = 'Homozygous'
-HEMI = 'Hemizygous'
-
-X_LINKED = 'X - linked'
-RECESSIVE = 'Autosomal recessive (homozygous)'
-DE_NOVO = 'de novo'
-DOMINANT = 'Autosomal dominant'
-
-
-def update_variant_inheritance(
-        variant_json: dict, family_individual_data: tuple[set[str], set[str], set[str], dict[str: list[str]]],
-        update_potential_comp_het_gene: Callable[str, None]) -> None:
-    """Compute the inheritance mode for the given variant and family"""
-
-    affected_individual_guids, unaffected_individual_guids, male_individual_guids, parent_guid_map = family_individual_data
-    is_x_linked = 'X' in variant_json.get('chrom', '')
-
-    genotype_zygosity = {
-        sample_guid: _get_genotype_zygosity(genotype, is_hemi_variant=is_x_linked and sample_guid in male_individual_guids)
-        for sample_guid, genotype in variant_json.get('genotypes', {}).items()
-    }
-    inheritance_model, possible_comp_het = _get_inheritance_model(
-        genotype_zygosity, affected_individual_guids, unaffected_individual_guids, parent_guid_map, is_x_linked)
-
-    if possible_comp_het:
-        for gene_id in variant_json.get('transcripts', {}).keys():
-            update_potential_comp_het_gene(gene_id)
-
-    variant_json.update({
-        'inheritance': inheritance_model,
-        'genotype_zygosity': genotype_zygosity,
-    })
-
-
-def _get_inheritance_model(
-        genotype_zygosity: dict[str, str], affected_individual_guids: set[str], unaffected_individual_guids: set[str],
-        parent_guid_map: dict[str: list[str]], is_x_linked: bool) -> tuple[str, bool]:
-
-    affected_zygosities = {genotype_zygosity[g] for g in affected_individual_guids if g in genotype_zygosity}
-    unaffected_zygosities = {genotype_zygosity[g] for g in unaffected_individual_guids if g in genotype_zygosity}
-
-    inheritance_model = ''
-    possible_comp_het = False
-    if any(zygosity in unaffected_zygosities for zygosity in {HOM_ALT, HEMI}):
-        # No valid inheritance modes for hom alt unaffected individuals
-        inheritance_model = ''
-    elif any(zygosity in affected_zygosities for zygosity in {HOM_ALT, HEMI}):
-        inheritance_model = X_LINKED if is_x_linked else RECESSIVE
-    elif HET in affected_zygosities:
-        if HET not in unaffected_zygosities:
-            inherited = (not unaffected_individual_guids) or any(
-                guid for guid in affected_individual_guids
-                if genotype_zygosity.get(guid) == HET and
-                any(genotype_zygosity.get(parent_guid) == HET for parent_guid in parent_guid_map[guid])
-            )
-            inheritance_model = DOMINANT if inherited else DE_NOVO
-
-        if len(unaffected_individual_guids) < 2 or HET in unaffected_zygosities:
-            possible_comp_het = True
-
-    return inheritance_model, possible_comp_het
-
-
-def _get_genotype_zygosity(genotype, is_hemi_variant):
-    num_alt = genotype.get('numAlt')
-    cn = genotype.get('cn')
-    if num_alt == 2 or cn == 0 or (cn != None and cn > 3):
-        return HOM_ALT
-    if num_alt == 1 or cn == 1 or cn == 3:
-        return HEMI if is_hemi_variant else HET
-    return None
-
-
-DISCOVERY_PHENOTYPE_CLASSES = {
-    'NEW': ['Tier 1 - Known gene, new phenotype', 'Tier 2 - Known gene, new phenotype'],
-    'EXPAN': ['Tier 1 - Phenotype expansion', 'Tier 1 - Novel mode of inheritance', 'Tier 2 - Phenotype expansion'],
-    'UE': ['Tier 1 - Phenotype not delineated', 'Tier 2 - Phenotype not delineated'],
-    'KNOWN': ['Known gene for phenotype'],
-}
-
-
-def get_discovery_phenotype_class(variant_tag_names):
-    for phenotype_class, class_tag_names in DISCOVERY_PHENOTYPE_CLASSES.items():
-        if any(tag in variant_tag_names for tag in class_tag_names):
-            return phenotype_class
-    return None
